+++ fluss-spark/fluss-spark-common/src/main/scala/org/apache/fluss/spark/read/FlussBatch.scala
@@ -160,7 +160,7 @@ class FlussAppendBatch(
     new FlussAppendPartitionReaderFactory(
       tablePath,
       projection,
-      pushedPredicate.orNull,
+      pushedPredicate,
       options,
       flussConfig)
   }
